Create functions for all buttons except operators
Numbers correctly display when pressed. AC button properly clears the
screen, displaying 0 when pressed. +/- converts the number to either
positive or negative. Decimal place is entered when pressed.

May be good to also find a way for the +/- button to be pressed and turn
the following number negative. Currently a number can only be made
negative after a positive number has been entered.

Will also need to create a way for the decimal button to be only capable
of being pressed once on a display, so that numbers do not have multiple
decimals in them.